[ESPN] Parse for more HTML info. in parsers
Parse for a bit more information in the HTML parsers in each
file. Output this information in the file dictionary getter
functions. This also removes the file_basename key as we can
now build each file output CSV name with the newly scraped data.
No need to be dependent on the input file name.