[STATSAPI] Add module to build local database
Added statsapi_db module that builds up a local database file
to store team and player information. This database of information
currently acts as a "local cache" to easily map between teams,
players and the links used to access the data from statsapi.

Implement the first function to create a database file. Creating
this database file simply creates a database with appropriate tables.
This is the starting point to build up this local database cache.
A database file must be created first before updating any entries
in the table with team and player information.