[STATSAPI] Add function to update players table
Added functionality to update players table in the database. Uses
an input CSV file containing player name, team (abbreviation), and
year string. Uses the team abbreviation to look up the team link
from the teams table. Then accesses the team roster json file from
statsapi for the given year string. Looks for the player on the
roster to retrieve name, ID, and statsapi link, and adds the entry
into the players table in the database.