Add ESPN to statsapi_db players interface script
Added a script that scrapes through all season folders for ESPN
draft and league roster information, and outputs a CSV of players
that is used as input for statsapi_db. This list of players should
contain all unique players from the draft and league rosters data.

TODO: It does not currently properly handle players with same names.