[ESPN-statsapi] Add validation script
Added initial ESPN to statsapi validation script. Purpose
is to ensure player names, teams, etc. are consistent
between ESPN and statsapi files. It does this by finding
a statsapi endpoint link for each player of interest from
the various ESPN parsed outputs.

Generates two outputs:
  1) A master list of all players
  2) An error log when a statsapi endpoint cannot be
     found for a player

Ideally, all errors are manually corrected early in the
inputs so handling of names and teams are not an issue for
downstream of the toolchain. The idea for now is to:
  1) Use this tool to check all players can be mapped
  2) If any errors, fix/correct this in the ESPN layer
  3) Repeat 1) and 2) until all errors are fixed (i.e.:
     all players have a valid statsapi endpoint mapping)

This script also exposes a few TODOs where it would be
convenient to have a file/data loading utility for ESPN
and statsapi data.