[ESPN-statsapi] Search suggestions in validation
Added an automatic search function in the ESPN-statsapi
validation script that attempts to look for the team of
a player of a given season. This function is only used
to help suggest corrections to the user in the error log.
This feature is not meant to help automatically correct
anything. A manual review to ensure errors and suggestions
are addressed is still required.

Also improved formatting of error output.