[ESPN-statsapi] Skip empty entries in validation
It is possible for ESPN parsed outputs to have a row where player
and team information is empty. Likely due to an empty slot on the
roster or something. Skip so it doesn't flag this as an error.

Commit also apparently checks in multiple line changes. Possibly
removing white spaces.