[ESPN] Add writer utility
Added a utility that writes various data to the various file
formats for the ESPN parsers. This utility is object-based
and holds a path to the output directory. Handles automatic
generation of the CSV, Excel, and pickles folders. Abstracts
logic required for data checking, Excel sheet writing, etc.
Using this utility is essentially the counter-part to the ESPN
loader utility.