[ESA] ESPN/statsapi analysis initial commit
Initial commit of the ESPN-statsapi analysis framework.
Framework loads in root ESPN and statsapi data folders
as supplied in the command-line arguments. Imports modules
from the "modules" folder and runs the "process()" function
for each module. Control which modules are ran using the
cfg.toml file. Added a couple of dummy files to start to
test the functionality of importing and running modules.