[ESPN] Add ESPN fantasy hockey web scraping module
Added class that scrapes the ESPN fantasy hockey league standings
and clubhouses website. Intended use of this module is to scrape
data from the league website to get day-by-day history. Use this
class in another script as needed. For example: Write a script
that calls into this one, and use Window's task scheduler to run
it periodically. Runner method scrapes league standings and each
team's clubhouse data. Saves the raw HTML to an archive folder.
Parses the HTML elements using ESPN HTML parser scripts to get
actual data of interest. This data is saved in a dictionary and
into a pickle to save storage space and make it easier to load.

Uses the Selenium module to open and control a web browser. For
now, this class uses Microsoft Edge since most Windows machines
will have Edge installed by default. This requires a web driver
executable that is compatible with the current browser version
you are using. Added drivers for Chrome, Firefox, and Edge into
the repo.

Note: For this script to work, league must be viewable by public.