[STATSAPI] Re-design downloader
Re-designed the statsapi downloader script to now be a class,
rather than a series of functions. This makes it more scalable
for adding new data to download. This also aims to make the
module much quicker and easier to use. Output folder structure
and file names have not changed and should be compatible with
the existing statsapi_loader.

New features include:
  - Using the async methods to download data much quicker
  - Downloading data is now mainly on a per-season basis
  - Removed the need for an input CSV with specific players;
    now downloads all players as long as they are on the roster
    for a given season

Also removed unit test for this module. Reason is because this
module simply utilizes the async functions from the utilities.
This module also auto-downloads a lot of files per method call,
so we don't want to send requests just for unit testing.