[ESA] Look for specific class in getmembers list
Updated logic to be more robust when looking for the correct
class to import given the module. This was needed because the
inspect.getmembers() function could return multiple classes
even if a file is simply using or importing classes from other
modules. Now, look for the class name that matches the module
name we want to dynamically import. Files and classes inside
the modules folder must follow the convention where file names
use underscores and the class inside it uses camel-case.

Example: example_stat.py will implement a class ExampleStat.