[ESA] Add wrapper utilities for plots
Added wrapper utilities for plotting pie and histograms. Allows
for using matplotlib or plotly backends by specifying a single
parameter.

Also updated matplotlib figsize to take in dimensions of the figures
in pixels, which gets converted back to inches in the utility. Updated
modules to use the wrapper utility and default to matplotlib backend
to start.