[ESPN/STATSAPI] Add data generator script
Source parsed and downloaded ESPN and statsapi data to generate
the raw data file. Meant to replace the data loading logic in ESA.
Generates files one time for downstream tools/scripts for ease
of analysis.

Majority of the logic was taken from the draft and standings stats
modules from ESA.