[ESPN] Scoring period parser checks 'stats' for GP
To handle a new edge case where the save stat for goalies appear to
be valid in the appliedStats dictionary even when they are not playing.
This causes an issue for the GP logic where they were credited with a
game played despite being on the bench.

Moving forward, a GP will only count if both appliedStats and stats
dictionaries are not empty.

Issue: EFHS-10