[ESPN] Parse for empty team name in player string
After manually editing the ESPN HTML files, there was a new case
where a player string in the draft recap file could contain an
empty team name. This is for players who were drafted but were not
on a valid team at the time.

Example: "Cody Glass , C" from 20182019 season.