[ESPN] Add new unified ESPN HTML parser class
Added a new general parser class that currently wraps draft recap and
league standings HTML parsing. The purpose of this is to replace and
simplify the current functionality of parsing HTML files. This class
can be used by other scripts to directly parse and obtain a dataframe
from the archived ESPN HTML files. This will clean-up and remove the
need to have a separate automation script to first pre-parse and output
data only to have it re-read and re-parsed again.

The only dependency here is that league standings files contain info
about team and owner names which draft recap does not have. This class
leverages the fact it parses both files and inserts owner name into the
parsed draft recap data. This will also replace needing to save
clubhouse files in the future (as that was originally the way to parse
for owner name given the team name).

Made changes in draft recap parser to not substitute special characters
in team names because we should not care about it matching file names
anymore. Also exposed a getter function in league standings parser to
obtain a dataframe mapping between team and owner names.